Update SELInux policy to refpol RELEASE_2_20240226
Pull in refpol release RELEASE_2_20240226 and use it to
replace the contents of packages/selinux-policy/selinux-policy.
This will be the starting point for RHEL 9 work.

NB: the policy will not build in this commit. That will be
fixed in subsequent commits. This commit is in isolation
to make it ieasier to identify and track the delta over
time.

Pulled from:
https://github.com/SELinuxProject/refpolicy/releases/tag/RELEASE_2_20240226

Unpacked form tarball:
https://github.com/SELinuxProject/refpolicy/releases/download/RELEASE_2_20240226/refpolicy-2.20240226.tar.bz2

with sha256sum:
7ed41f4f45189b9ee9706da8ac357eccc103651b56daabaddb54c436e8117cf9

The modifications in this commit are copyrighted works of Cubic Corporation.
(C) 2024 Cubic Corporation