let systemd-generator create units
systemd-generator was missing rules allowing it to
dynamically create units, add them

This is a rebaseed cherry-pick of a systemd generator
fix introduced in 9e8768cc7733216741f55ac87b8a8d315888b4a2
for the RHEL 8 policy.

The modifications in this commit are copyrighted works of Cubic Corporation.
(C) 2024 Cubic Corporation